+++ config/config.exs
@@ -42,7 +42,7 @@ config :esbuild,
     args:
       ~w(js/app.js
          js/workflow-diagram/component.tsx
-         js/compiler/index.ts
+         js/adaptor-docs/index.ts
          js/editor/Editor.tsx
          --format=esm --splitting --bundle
          --target=es2020