+++ lib/lightning_web/live/dataclip_live/edit.ex
@@ -1,7 +1,6 @@
 defmodule LightningWeb.DataclipLive.Edit do
   @moduledoc """
-  LiveView for editing a single dataclip, which inturn uses
-  `LightningWeb.JobLive.JobFormComponent` for common functionality.
+  LiveView for editing a single dataclip.
   """
   use LightningWeb, :live_view