+++ lib/lightning_web/live/live_helpers.ex
@@ -27,11 +27,17 @@ defmodule LightningWeb.LiveHelpers do
   """
 
   def live_info_block(assigns) do
-    assigns = assign_new(assigns, :myself, fn -> nil end)
+    assigns =
+      assigns
+      |> assign_new(:myself, fn -> nil end)
+      |> assign_new(:id, fn -> Ecto.UUID.generate() end)
 
     ~H"""
     <%= if info = live_flash(@flash, :info) do %>
-      <div class="fixed flex justify-center bottom-3 right-0 left-0 z-[100]">
+      <div
+        class="fixed w-fit mx-auto flex justify-center bottom-3 right-0 left-0 z-[100]"
+        id={@id}
+      >
         <p
           class="bg-blue-200 border-blue-300 border opacity-75 py-4 px-5 rounded-md drop-shadow-lg"
           role="alert"
@@ -47,11 +53,17 @@ defmodule LightningWeb.LiveHelpers do
   end
 
   def live_error_block(assigns) do
-    assigns = assign_new(assigns, :myself, fn -> nil end)
+    assigns =
+      assigns
+      |> assign_new(:myself, fn -> nil end)
+      |> assign_new(:id, fn -> Ecto.UUID.generate() end)
 
     ~H"""
     <%= if error = live_flash(@flash, :error) do %>
-      <div class="fixed flex justify-center bottom-3 right-0 left-0 z-[100]">
+      <div
+        class="fixed w-fit mx-auto flex justify-center bottom-3 right-0 left-0 z-[100]"
+        id={@id}
+      >
         <p
           class="bg-red-300 border-red-400 text-red-900 border opacity-75 py-4 px-5 rounded-md drop-shadow-lg"
           role="alert"